requirements.txt
streamlit
feedparser
beautifulsoup4
requests
python-dateutil
newspaper3k
lxml[html_clean]
googlenewsdecoder
pyngrok
